Auto-issue on Dependabot fail
Trying out a new (and exciting?) way to report build issues when Dependabot bumps, and automatically open a ticket.

The reasoning is that usually, when Dependabot fails to build the ruleset against the latest version of osu! there must be breaking changes. We can then open a ticket for the failed PR and attach the logs for organization.